change weight values to int and how powerModifier works
Pokemon's weight value is now constructed as an int value and is in hectograms (kilograms * 10).
CalculateLowKickPower now does int comparisons. Also cleaned up that method a bit.
There is now a move power modifier in BattleContext set by a move routine (for now). CalculateDamage uses it to set currentMovePower and powerModifier is taken out of the main baseDamage equation.