11/8/2025 some changes
Changed MoveRoutines to use std::variant rather than runtime polymorphism. Less verbose, and (very slightly) faster (AI only battles seemed snappier). Working on implementing a move scoring type system for the AI similar to the official games (not exact). Not close to being done yet. (Not in this upload but some additions to AIPlayer, BattleManager, and an addition of a BattleAIProcedures namespace).